refactor(services): extract web session query and runner modules
Eleven read-side methods move into web-session-query-client.ts over the
new SessionQueryService: the workflow trio, getSessionDetails, the
session list/search/get pair, getActiveSession, exportSession and
listAgentRunners, with searchText, sessionSearchMatches and
serializeWebSessionExport moving alongside as their private helpers.

The runner descriptor helpers move to web-agent-runner.ts because
sendMessage and listAgentRunners now sit in different modules;
sendMessage keeps reaching them through selectWebRunner and
webRunRunner.

Signatures moved out of AgentService rather than being copied, which is
why agent-service.ts drops another 18 lines.

web-agent-client.ts 1332 -> 1105, agent-service.ts 338 -> 320.
Subtree aggregate 19286 -> 19319.

tsc --noEmit, contracts:check and the full suite pass: 286 files /
1301 tests, unchanged, with no test file edited.
